requirements.txt
absl-py==1.4.0
aiohttp==3.8.6
aiosignal==1.3.1
albumentations==1.3.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.1.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1687772187254/work
bpemb==0.3.4
cachetools==5.3.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer==3.0.1
clip==1.0
colorama @ file:///croot/colorama_1672386526460/work
cycler==0.11.0
Cython @ file:///croot/cython_1676568029361/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
DNN-printer==0.0.2
efficientnet-pytorch==0.7.1
einops==0.6.1
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649908585034/work
et-xmlfile==1.1.0
filelock==3.12.2
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
fvcore==0.1.5.post20221221
gensim==4.2.0
google-auth==2.17.3
google-auth-oauthlib==0.4.6
grpcio==1.54.0
h5py==3.8.0
huggingface-hub==0.16.4
idna==3.4
ignite==1.1.0
imageio==2.25.1
imgaug==0.4.0
importlib-metadata==6.6.0
install==1.3.5
iopath==0.1.10
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1620912939357/work/dist/ipykernel-5.5.5-py3-none-any.whl
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1651240553635/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
joblib==1.2.0
jupyter-client==5.3.4
jupyter-core==4.5.0
kiwisolver==1.4.4
libsvm==3.14
lightning-utilities==0.10.1
lpips==0.1.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
multidict==6.0.5
networkx==2.6.3
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.7.0.68
opencv-python-headless==4.7.0.72
openpyxl==3.0.10
packaging==23.0
pandas==1.3.5
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.4.0
portalocker @ file:///tmp/build/80754af9/portalocker_1617135539189/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1688565951714/work
protobuf==3.20.3
ptflops==0.6.9
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybrisque==1.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1691408637400/work
pyignite==0.6.1
pyparsing==3.0.9
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytorch-ignite==0.4.12
pytorch-lightning==1.9.5
pytz==2023.3
PyWavelets==1.3.0
PyYAML==6.0
pyzmq @ file:///opt/conda/conda-bld/pyzmq_1657724186960/work
qudida==0.0.4
regex @ file:///tmp/abs_41f5bce5-0a2e-45aa-b231-1fd2fbd57753gfpe6sjm/croots/recipe/regex_1658257178822/work
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
sacrebleu @ file:///tmp/build/80754af9/sacrebleu_1634578602645/work
safetensors==0.3.1
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
sentencepiece==0.1.97
shapely==2.0.3
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==6.3.0
svm==0.1.0
tabulate @ file:///opt/conda/conda-bld/tabulate_1657784105888/work
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-logger==0.1.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
termcolor==2.3.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.9.2
torch==1.10.1+cu113
torchaudio==0.10.1+cu113
torchinfo==1.8.0
torchmetrics==0.11.4
torchsummary==1.5.1
torchvision==0.11.2+cu113
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827244717/work
tqdm==4.64.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1675110562325/work
typing_extensions==4.5.0
urllib3==1.26.14
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1696255154857/work
Werkzeug==2.2.3
wrapt==1.15.0
xmltodict==0.13.0
yacs==0.1.8
yarl==1.9.4
zipp==3.15.0
